Revert "Let everything in tmp directory be owned by app user"
This reverts commit d77f2ca19fa5b6854d16cb0b6b46ba5939d89d63.

This was originally done because the attachment_fu directory in tmp was owned
by root and that was causing deleting files to fail. However, that directory
is not created actually created during deploy, but when Rails boots. So it
actually is created by running migrations (which happens as the root user
later). So the actual solution will be to not use cap deploy:migrations and
run migrations by hand.